libs/utils/analysis: add EAS module
This patch move the EAS/SchedTune analysis code, originally part of the
trace_analysis.py module, to be a new TraceAnalysis module with dynamic
registration via the AnalysisRegister.
